server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
line-bot-sdk==3.22.0
apscheduler==3.10.4
